Restructure python packages
