+++ ansible_collections/arista/avd/roles/eos_designs/defaults/main/default-templates.yml
@@ -5,7 +5,7 @@ default_templates:
     structured_config:
       # Render Structured Configuration
       # Base features
-      - python_module: "ansible_collections.arista.avd.roles.eos_designs.python_modules.base.base"
+      - python_module: "ansible_collections.arista.avd.roles.eos_designs.python_modules.base"
         python_class_name: "AvdStructuredConfig"
       # MLAG feature
       - python_module: "ansible_collections.arista.avd.roles.eos_designs.python_modules.mlag"
@@ -46,7 +46,7 @@ default_templates:
   mpls:
     structured_config:
       # Render Structured Configuration
-      - python_module: "ansible_collections.arista.avd.roles.eos_designs.python_modules.base.base"
+      - python_module: "ansible_collections.arista.avd.roles.eos_designs.python_modules.base"
         python_class_name: "AvdStructuredConfig"
       # Underlay feature
       - template: "underlay/main.j2"
@@ -82,7 +82,7 @@ default_templates:
     structured_config:
       # Render Structured Configuration
       # Base features
-      - python_module: "ansible_collections.arista.avd.roles.eos_designs.python_modules.base.base"
+      - python_module: "ansible_collections.arista.avd.roles.eos_designs.python_modules.base"
         python_class_name: "AvdStructuredConfig"
       # MLAG feature
       - python_module: "ansible_collections.arista.avd.roles.eos_designs.python_modules.mlag"
