+++ ansible_collections/arista/avd/roles/eos_designs/python_modules/network_services/ethernet_interfaces.py
@@ -5,7 +5,8 @@ from functools import cached_property
 
 from ansible_collections.arista.avd.plugins.filter.natural_sort import natural_sort
 from ansible_collections.arista.avd.plugins.plugin_utils.utils import get
-from ansible_collections.arista.avd.roles.eos_designs.python_modules.network_services.utils import UtilsMixin
+
+from .utils import UtilsMixin
 
 
 class EthernetInterfacesMixin(UtilsMixin):
