+++ ansible_collections/arista/avd/roles/eos_designs/python_modules/network_services/ip_igmp_snooping.py
@@ -3,7 +3,8 @@ from __future__ import annotations
 from functools import cached_property
 
 from ansible_collections.arista.avd.plugins.plugin_utils.utils import default, get
-from ansible_collections.arista.avd.roles.eos_designs.python_modules.network_services.utils import UtilsMixin
+
+from .utils import UtilsMixin
 
 
 class IpIgmpSnoopingMixin(UtilsMixin):
