+++ ansible_collections/arista/avd/roles/eos_designs/python_modules/network_services/ipv6_static_routes.py
@@ -2,7 +2,7 @@ from __future__ import annotations
 
 from functools import cached_property
 
-from ansible_collections.arista.avd.roles.eos_designs.python_modules.network_services.utils import UtilsMixin
+from .utils import UtilsMixin
 
 
 class Ipv6StaticRoutesMixin(UtilsMixin):
