+++ ansible_collections/arista/avd/roles/eos_designs/python_modules/network_services/port_channel_interfaces.py
@@ -6,7 +6,8 @@ from functools import cached_property
 from ansible_collections.arista.avd.plugins.filter.esi_management import generate_esi, generate_lacp_id, generate_route_target
 from ansible_collections.arista.avd.plugins.filter.natural_sort import natural_sort
 from ansible_collections.arista.avd.plugins.plugin_utils.utils import get
-from ansible_collections.arista.avd.roles.eos_designs.python_modules.network_services.utils import UtilsMixin
+
+from .utils import UtilsMixin
 
 
 class PortChannelInterfacesMixin(UtilsMixin):
