+++ ansible_collections/arista/avd/roles/eos_designs/python_modules/network_services/struct_cfg.py
@@ -4,7 +4,8 @@ from functools import cached_property
 
 from ansible_collections.arista.avd.plugins.plugin_utils.merge import merge
 from ansible_collections.arista.avd.plugins.plugin_utils.utils import get
-from ansible_collections.arista.avd.roles.eos_designs.python_modules.network_services.utils import UtilsMixin
+
+from .utils import UtilsMixin
 
 
 class StructCfgMixin(UtilsMixin):
